Adding hook for keybinding registration Adding migration for settings store Moving DependencyInjection Configuration back to Core
