+++ bundles/AdminBundle/public/css/icons.css
@@ -726,10 +726,6 @@
     background: url(/bundles/pimcoreadmin/img/flat-color-icons/filing_cabinet.svg) center center no-repeat !important;
 }
 
-.pimcore_icon_glossary {
-    background: url(/bundles/pimcoreadmin/img/flat-color-icons/book.svg) center center no-repeat !important;
-}
-
 .pimcore_icon_tabs {
     background: url(/bundles/pimcoreadmin/img/flat-color-icons/tabs.svg) center center no-repeat !important;
 }
