+++ bundles/GlossaryBundle/config/pimcore/config.yaml
@@ -1,0 +1,3 @@
+imports:
+    - { resource: default.yaml }
+    - { resource: doctrine_migrations.yaml }
