+++ bundles/GlossaryBundle/config/pimcore/default.yaml
