+++ bundles/GlossaryBundle/config/pimcore/doctrine_migrations.yaml
@@ -1,0 +1,3 @@
+doctrine_migrations:
+    migrations_paths:
+        'Pimcore\Bundle\GlossaryBundle\Migrations': '@GlossaryBundle/src/Migrations'
