+++ bundles/GlossaryBundle/public/css/glossary.css
@@ -1,0 +1,3 @@
+.pimcore_icon_glossary {
+    background: url(/bundles/pimcoreadmin/img/flat-color-icons/book.svg) center center no-repeat !important;
+}
