--- bundles/GlossaryBundle/src/DependencyInjection/Configuration.php
@@ -1,50 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * Pimcore
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Pimcore Commercial License (PCL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- *  @license    http://www.pimcore.org/license     GPLv3 and PCL
- */
-
-namespace Pimcore\Bundle\GlossaryBundle\DependencyInjection;
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Builder\TreeBuilder;
-use Symfony\Component\Config\Definition\ConfigurationInterface;
-
-/**
- * @internal
- */
-final class Configuration implements ConfigurationInterface
-{
-
-    public function getConfigTreeBuilder()
-    {
-        $treeBuilder = new TreeBuilder('pimcore_glossary');
-        /** @var ArrayNodeDefinition $rootNode */
-        $rootNode = $treeBuilder->getRootNode();
-        $rootNode->addDefaultsIfNotSet();
-
-        $this->addGlossaryNode($rootNode);
-        return $treeBuilder;
-    }
-
-    private function addGlossaryNode(ArrayNodeDefinition $rootNode): void
-    {
-        $rootNode
-            ->children()
-            ->arrayNode('glossary')
-            ->addDefaultsIfNotSet()
-            ->children()
-            ->arrayNode('blocked_tags')
-            ->useAttributeAsKey('name')
-            ->prototype('scalar');
-    }
-}
