+++ bundles/GlossaryBundle/src/DependencyInjection/GlossaryExtension.php
@@ -19,15 +19,14 @@ namespace Pimcore\Bundle\GlossaryBundle\DependencyInjection;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
+use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\ConfigurableExtension;
 
-final class GlossaryExtension extends ConfigurableExtension
+
+final class GlossaryExtension extends Extension implements PrependExtensionInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function loadInternal(array $config, ContainerBuilder $container): void
+    public function load(array $configs, ContainerBuilder $container)
     {
         // on container build the shutdown handler shouldn't be called
         // for details please see https://github.com/pimcore/pimcore/issues/4709
@@ -38,15 +37,11 @@ final class GlossaryExtension extends ConfigurableExtension
             new FileLocator(__DIR__ . '/../../config')
         );
 
-        //$loader->load('default.yaml');
         $loader->load('services.yaml');
-
-
-        $this->configureGlossary($container, $config['glossary']);
     }
 
-    private function configureGlossary(ContainerBuilder $container, array $config): void
+    public function prepend(ContainerBuilder $container): void
     {
-        $container->setParameter('pimcore.glossary.blocked_tags', $config['blocked_tags']);
+        // doctrine_migrations.yaml loaded with pimcore config.yaml as in CoreBundle
     }
 }
